Add files via upload
This is my personal portfolio website built using HTML, CSS, and JavaScript. It showcases my projects, skills, and contact information in a clean and responsive design. The website is designed to highlight my work as a beginner web developer and serves as an online resume for potential employers and clients.